+++ src/JointFiles.kt
@@ -1,6 +1,5 @@
 import java.io.OutputStreamWriter
 import kotlin.math.log10
-import org.apache.hadoop.fs.FileSystem
 import org.apache.hadoop.fs.Path
 import org.apache.log4j.LogManager
 import org.apache.spark.api.java.JavaSparkContext
@@ -13,7 +12,6 @@ import org.jetbrains.kotlinx.spark.api.*
 // (ii) the genomic ranges covered by the remaining parts.
 // For parts (ii): rename them so that filename indicates the genomic range
 // For parts (i): split them by CHROM, naming each subpart according to the same schema
-
 fun reorgJointFiles(spark: SparkSession, pvcfDir: String, aggHeader: AggVcfHeader) {
     val logger = LogManager.getLogger("vcfGLuer")
     val jsc = JavaSparkContext(spark.sparkContext)
@@ -36,7 +34,7 @@ fun reorgJointFiles(spark: SparkSession, pvcfDir: String, aggHeader: AggVcfHeade
             basenames.asSequence().map {
                     basename ->
                 fileReaderDetectGz("$pvcfDir/$basename", fs).useLines {
-                    val rec = parseVcfRecord(aggHeaderB.value.contigId, -1, it.first())
+                    val rec = parseVcfRecord(aggHeaderB.value.contigId, it.first())
                     basename to rec.range
                 }
             }.iterator()
@@ -131,7 +129,7 @@ fun splitByChr(
     fileReaderDetectGz("$pvcfDir/$partBasename", fs).useLines {
         var writer: OutputStreamWriter? = null
         for (line in it) {
-            val rec = parseVcfRecord(aggHeader.contigId, -1, line)
+            val rec = parseVcfRecord(aggHeader.contigId, line)
             if (rec.range.rid != rid) {
                 if (writer != null) {
                     writer.close()
@@ -161,16 +159,4 @@ fun splitByChr(
     check(fs.delete(Path(pvcfDir, partBasename), false), { "unable to delete $partBasename" })
 
     return ans
-}
-
-fun FileSystem.listSequence(
-    path: String,
-    recursive: Boolean
-): Sequence<org.apache.hadoop.fs.LocatedFileStatus> {
-    val it = getFileSystem(path).listFiles(Path(path), recursive)
-    return sequence {
-        while (it.hasNext()) {
-            yield(it.next())
-        }
-    }
 }
